Console: Stop `tasks` from hanging in `std::sort`.
`std::sort` expects a comparator that returns `a < b`. Flipping this to `a >= b`
would normally be fine to reverse the order, but floats behave weirdly with NaN.

Instead of flipping the comparator, this uses the reverse-iterators to reverse
the sort order of the tasks, and returns to an `a < b` comparator.