Include more character sets.
This adds support for Latin Extended-A, Latin Extended-B, and Greek
and Coptic scripts. The used fusion font in the current version
support for these is clearly incomplete for the first two, there is
however an upstream branch of the ark pixel font that is a base for it
to add more support for those.

In the current state this reduces the amount of missing characters for
a whole set of Eastern and South-Eastern European languages.