rework the lua i2c api to return multiple values instead of a table
This gets us out of having to store strings, which means the command objects become trivially destructible (now asserted at compile time).

Also run script reloads on a worker task and not the main task. Apparently I was *very* close to overflowing the main task's stack and happened to change something in the script in just the right way to push it over the limit.