Add ark pixel font's latin feature branch font.
This is from an upstream development branch at [1] and adds quite a lot
of glyps for the 12px font, and a few to the 10px font.

This addition should be temporary until that feature branch is merged
upstream, a new release rolled, and then a new release of the fuxion
pixel font is rolled. For now this leads to no more missing extended
latin A/B characters in the 12px font on my collection.

[1] https://github.com/TakWolf/ark-pixel-font/commit/535181d3c9394990fc0cd5d5259399da3b7ead48